+++ packages/example-backend-core/src/errors/UserDoesNotHaveRoleErrorInThisScope.error.ts
@@ -1,0 +1,9 @@
+class UserDoesNotHaveRoleErrorInThisScopeError extends Error {
+  name = "UserDoesNotHaveRoleErrorInThisScopeError";
+
+  constructor() {
+    super("The user does not have the role in the scope!");
+  }
+}
+
+export { UserDoesNotHaveRoleErrorInThisScopeError };