+++ packages/example-backend-core/src/tests/assignScopeToUser.test.ts
@@ -1,5 +1,11 @@
 import { ScopeController, UserController } from "../mocks";
-import { assignScopeToUser } from "../utils";
+import {
+  assignScopeToUser,
+  getScope,
+  getUser,
+  updateScope,
+  updateUser
+} from "../utils";
 
 it("Should assign a scope to a user", async () => {
   const scope = new ScopeController("scope1");
@@ -7,10 +13,12 @@ it("Should assign a scope to a user", async () => {
   const scopeWithUser = await assignScopeToUser(
     "scope1234",
     "user1234",
-    id => scope.get(id),
-    id => user.get(id),
-    (id, newScope) => scope.update(id, newScope),
-    (id, newUser) => user.update(id, newUser)
+    id => getScope(id, id => scope.get(id)),
+    id => getUser(id, id => user.get(id)),
+    (id, newScope) =>
+      updateScope(id, newScope, (id, newScope) => scope.update(id, newScope)),
+    (id, newUser) =>
+      updateUser(id, newUser, (id, newUser) => user.update(id, newUser))
   );
   expect((await scopeWithUser).users[0]).toMatchObject({
     name: "user1",